Changed how children template is set up.
Instead of being an object with types as a key and a range as a value, the child template is now an array of objects which include the type and range as variables. This allows adding the same type of child multiple times to achieve a non-linear distribution. This also paves the way for adding validation in the future (ie, only letting objects with certain attributes have certain children).